moving style wrap to email container so original email is not modified
